Fix defects in forwarding, verification, scoring inputs and mix planning
All of these were reproduced against this checkout, not in isolation.

CLI forwarding (the documented train command could not start)
  `train sft` declares --budget-minutes as float, _forward stringifies it
  to "90.0", and the target script parsed it as int -- so the README
  command exited 2 before any work began, even in dry-run. The three
  train scripts now take float. tests/test_cli.py replaces cli._script
  with a stub and tests/test_docs_commands.py stops at build_parser(),
  so neither crossed the boundary where this breaks; the new
  tests/test_cli_forwarding.py walks the real chain to the trainer.

One definition of a correct numeric answer (src/boldt_posttrain/verifiers.py)
  The evaluation validator required the whole trimmed output to be the
  number, the RL reward took the last number anywhere, and failure-mining
  synthesis took the first. For expected 42, "20 + 22 = 42" was rejected
  by synthesis but rewarded by RL, and "42 ist falsch. Das Ergebnis ist
  7." the other way round -- and synthesis turns those verdicts into
  chosen/rejected preference pairs. All three now share the evaluation
  rule, because that is what decides promotion.

Ordered expectations are checked in order
  _default_validate accepted "Beta Alpha" for ["Alpha", "Beta"] although
  the cases say "in dieser Reihenfolge".

Leakage and license are derived, not asserted
  run_real_evaluation set leakage=clean and license=reviewed_usable
  unconditionally, and the scorer uses both as hard gates. They now come
  from manifest.json, leakage_report.json and discovery.json, tied to the
  decontamination hash; anything unreadable or stale is reported as
  not_checked/unknown, which the scorer already fails closed on.

Model revision reaches the lm-eval subprocess
  Local generation loaded the pinned revision while lm_eval got only
  pretrained=..., so the two halves of one evaluation could score
  different weights.

coding is mapped to a metric
  It is a scored category in the suite but was missing from the
  category-to-metric map, so its cases never reached the metrics.

build_mix_plan normalizes correctly
  Enforcing minimum weights only reduced groups without a minimum, so
  safety=0.9/language=0.1 with floors 0.1/0.2 raised "no feasible
  normalized mix" although 0.8/0.2 satisfies both. Groups may now give up
  the share above their own floor. Required groups that no probe covers
  were silently dropped, which let a pure "general" mix report status ok
  while the safety and language minimums went unenforced; that now fails.

PiSSA adapters reload to the trained model
  PiSSA rewrites the base weights at initialisation, so a plain
  save_pretrained produced an adapter that no longer matched the base it
  was reloaded onto: measured max|delta logits| 0.0397 on the CPU fixture
  versus exactly 0 for LoRA and rsLoRA. The existing test asserted only
  that the logits were finite, which both cases satisfy. Training now
  captures the pre-training adapter at on_train_begin and converts on
  save (max|delta| 1.5e-08), and the test asserts equivalence. The
  capture must come from the same model instance -- pissa_niter_4 uses a
  randomized SVD, so a re-initialised model is a different decomposition.
  Preference training rejects PiSSA instead: its trainer is built and run
  inside train_preference, so there is no point before the first step to
  capture from. PiSSA also requires an unquantized base, which makes it
  unreachable under a qlora config, and "method" is not a search override
  dimension -- configs/posttrain/experiments/search-example.json inherited
  qlora, so its third trial could never have run; it now pins lora.

Verified: ruff check, ruff format --check, full suite on an RTX A6000.

+++ scripts/pt_train_preference.py
@@ -31,7 +31,9 @@ def main(argv: Optional[List[str]] = None) -> int:
     ap.add_argument("--specialist", default="preference-de")
     ap.add_argument("--out", default=str(ROOT / "outputs/posttrain/runs"))
     ap.add_argument("--data", default=str(ROOT / "outputs/posttrain/data"))
-    ap.add_argument("--budget-minutes", type=int, default=90)
+    # float, not int: the CLI declares this as float and forwards str(value),
+    # so an int parser here rejects the documented "--budget-minutes 90" as "90.0".
+    ap.add_argument("--budget-minutes", type=float, default=90)
     ap.add_argument("--dry-run", action="store_true")
     ap.add_argument("--real", action="store_true")
     ap.add_argument("--allow-gpu", action="store_true")
